+++ tests/unix_socket.rs
@@ -1,8 +1,8 @@
 use anyhow::Result;
-use async_std::task;
 use pretty_assertions::assert_eq;
 use serde_cbor::de::from_slice;
 use serde_cbor::ser::to_vec;
+use tokio::task;
 
 use pueue_lib::network::message::*;
 use pueue_lib::network::protocol::*;
@@ -10,7 +10,7 @@ use pueue_lib::network::protocol::*;
 mod helper;
 
 #[cfg(not(target_os = "windows"))]
-#[async_std::test]
+#[tokio::test]
 /// This tests whether we can create a listener and client, that communicate via unix sockets.
 async fn test_unix_socket() -> Result<()> {
     better_panic::install();